# openapi-2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lunch Guide API",
    "description": "This is the API for the Lunch Guide made by Ael Banyard, Evan Lüber and Yannick Schläpfer.",
    "termsOfService": "http://lunch-guide.ch/agb/",
    "contact": {
      "email": "api@lunch-guide.ch"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.1"
  },
...
  "servers": [
    {
      "url": "https://lunch-guide.ch/api/"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Login to API",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "User not verified"
          },
          "402": {
            "description": "Password wrong"
          },
          "403": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "name": "token",
        "in": "header"
      }
    }
  }
}